services/fastapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.3.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
kafka-python==2.0.2
mlflow==2.13.0
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.0
